backup/requirements.txt
Flask
requests
google-oauth-lib
google-auth-oauthlib
google-api-python-client
langchain
google-cloud-aiplatform
pandas
pdfplumber
Pillow==10.3.0
vertexai
langchain-google-vertexai
httpx==0.28.1